requirements.txt
# local package
-e .

# external requirements
click~=8.1.7
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=2.0.3
torch~=2.1.2
numpy~=1.24.4
matplotlib~=3.7.5
torchmetrics~=1.3.2
torchvision~=0.16.2
Pillow~=10.1.0
setuptools~=68.0.0
src~=0.1.0
seaborn~=0.12.2
scikit-learn~=1.3.2